Introduced a base class for SVO's based on a string.
